requirements.txt
flask
anthropic
Pillow
python-dotenv
requests
